+++ Kernel/System/ProcessManagement/TransitionAction/Base.pm
@@ -349,7 +349,16 @@ sub _ReplaceTicketAttributes {
 
     # if we have a TicketArticleCreate transition action then we can't use the data of the parameters
     # for the template generator
-    my $ArticleHashRef = {};
+    #
+    # ContentType/Attachment (added below when resolving <OTRS_FIRST_ARTICLE_Body>/<OTRS_LAST_ARTICLE_Body>)
+    # are only meaningful for transition actions that actually build an article Body; for any other
+    # caller (e.g. DynamicFieldSet) they must not be added to Config, or they get mistaken for
+    # (missing) dynamic field names, even if the config attribute they were resolved for happens
+    # to be named 'Body' too
+    # if a new transition action module is added that also builds an article Body and needs
+    # ContentType/Attachment, add its package name to the regex below as well
+    my $ArticleHashRef      = {};
+    my $IsArticleBodyCaller = 0;
     if ( IsHashRefWithData( $Param{Config} ) ) {
         CALLER:
         for my $Caller ( 0 .. 10 ) {
@@ -357,9 +366,13 @@ sub _ReplaceTicketAttributes {
 
             last CALLER if !$Package1;
 
-            next CALLER if $Subroutine1 !~ m{ ProcessManagement\:\:TransitionAction\:\:TicketArticleCreate }xmsi;
+            next CALLER
+                if $Subroutine1
+                !~ m{ ProcessManagement\:\:TransitionAction\:\:(?:ArticleSend|TicketCreate|TicketArticleCreate) }xmsi;
 
-            $ArticleHashRef = $Param{Config};
+            $IsArticleBodyCaller = 1;
+            $ArticleHashRef      = $Param{Config}
+                if $Subroutine1 =~ m{ ProcessManagement\:\:TransitionAction\:\:TicketArticleCreate }xmsi;
 
             last CALLER;
         }
@@ -593,13 +606,20 @@ sub _ReplaceTicketAttributes {
                                 String => $Article{Body},
                             );
                         }
-                        $Param{Config}->{ContentType} = 'text/html; charset="utf-8"';
+
+                        # only pollute Config with ContentType/Attachment when the tag was resolved
+                        # for the actual article Body of a transition action that builds one, not
+                        # e.g. for an unrelated DynamicFieldSet config key (see bug report about
+                        # Attachment/ContentType leaking in and being mistaken for dynamic field names)
+                        if ( $Attribute eq 'Body' && $IsArticleBodyCaller ) {
+                            $Param{Config}->{ContentType} = 'text/html; charset="utf-8"';
+
+                            # get all attachments if there is more than one article tag
+                            $Param{Config}->{Attachment} ||= [];
+                            push @{ $Param{Config}->{Attachment} }, @{ $ArticleAttachments{$ArticleType} };
+                        }
                         $Value = "<blockquoute>$Value</blockquoute>";
                         Encode::_utf8_on($Value);
-
-                        # get all attachments if there is more than one article tag
-                        $Param{Config}->{Attachment} ||= [];
-                        push @{ $Param{Config}->{Attachment} }, @{ $ArticleAttachments{$ArticleType} };
                     }
 
                     # replace tag in data